Table creation script + remove accidental file
These tables aren't necesarily the best, but they're the ones that I
inherited this database with.

The other file is one that I accidentally created & committed due to
MySQL Workbench not saving files with a `.sql` unless you specify it, so
it didn't show up and I resaved it under a new name.